users/serializers.py
from users.models import Person
from rest_framework import serializers


class PersonSerializer(serializers.ModelSerializer):
    class Meta:
        model = Person
        fields = (
            'id',
            'description',
            'balance',
            'username',
            'password'
        )
